Tidied up UserRepository and removed a memory leak from DataSource.
