+++ app/src/main/java/com/thinqtv/thinqtv_android/ui/login/LoginViewModelFactory.java
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel;
 import androidx.lifecycle.ViewModelProvider;
 import androidx.annotation.NonNull;
 
-import com.thinqtv.thinqtv_android.data.LoginRepository;
+import com.thinqtv.thinqtv_android.data.UserRepository;
 
 /**
  * ViewModel provider factory to instantiate LoginViewModel.
@@ -17,10 +17,10 @@ public class LoginViewModelFactory implements ViewModelProvider.Factory {
     @SuppressWarnings("unchecked")
     public <T extends ViewModel> T create(@NonNull Class<T> modelClass) {
         if (modelClass.isAssignableFrom(LoginViewModel.class)) {
-            return (T) new LoginViewModel(LoginRepository.getInstance());
+            return (T) new LoginViewModel(UserRepository.getInstance());
         }
         else if (modelClass.isAssignableFrom(RegisterViewModel.class)) {
-            return (T) new RegisterViewModel(LoginRepository.getInstance());
+            return (T) new RegisterViewModel(UserRepository.getInstance());
         }
         else {
             throw new IllegalArgumentException("Unknown ViewModel class");
